Fix float/double type conversion and remove incorrect static import
- Fix CoveragePenaltyOccurredSkillsCheck: Cast float Odds to double before multiplication
- Remove incorrect static import in PenaltySkillsChecksTests.cs (TackleContext is in namespace, not nested class)

All compilation errors resolved.

+++ StateLibrary/SkillsChecks/CoveragePenaltyOccurredSkillsCheck.cs
@@ -138,7 +138,7 @@ namespace StateLibrary.SkillsChecks
             foreach (var penaltyName in penalties)
             {
                 var basePenalty = Penalties.List.Single(p => p.Name == penaltyName);
-                var baseOdds = basePenalty.Odds;
+                var baseOdds = (double)basePenalty.Odds; // Convert float to double
 
                 // Context-specific adjustments
                 if (penaltyName == PenaltyNames.DefensivePassInterference)
